Further bug fixes
