+++ src/MangaUpdates/includes/icon.png
